requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
packaging>=21.0
pycocotools>=2.0.6
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
roboflow>=1.1.0
opencv-python>=4.8.0
timm>=0.9.0
scipy>=1.10.0

